requirements.txt
# FastAPI BigQuery Service - Phase 1 (MVP)

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# BigQuery
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-auth==2.35.0

# Claude Code CLI (via npm wrapper)
anthropic==0.40.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12

# Monitoring & Logging
loguru==0.7.2

# HTTP Client
httpx==0.27.2

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1

# Security Testing
bandit==1.7.9
safety==3.2.7
pip-audit==2.7.3
